+++ src/Ext/Techno/Hooks.cpp
@@ -251,7 +251,7 @@ DEFINE_HOOK(0x71067B, TechnoClass_EnterTransport_LaserTrails, 0x7)
 }
 
 // I don't think buildings should have laser-trails
-DEFINE_HOOK(0x4D7221, FootClass_Unlimbo_LaserTrails, 0x6)
+DEFINE_HOOK(0x4D7221, FootClass_Unlimbo, 0x6)
 {
 	GET(FootClass*, pTechno, ESI);